Optional error_key filter for command inputs
Description: There are cases in which the keys used in a command do not
fully describe the interface by which the command receives its inputs.
This can cause cases where validation errors returned in a command
don't make as much sense. For example, a command which takes an input
of:

```ruby
hash :company do
  required do
    string :name
  end
end
```

would, if blank, return the error "Name cannot be blank", when a more
suitable error message may look something like "Company Name cannot
be blank".

This commit adds an optional field which
will allows one display error messages using an optional error key
which, when present, will override the key passed in:

```ruby
hash :company do
  required do
    string :name, error_key: :company_name
  end
end
```

Which will return the error message we desire.