Respect Mutations.cache_constants in array filters
Model and array filters both have a `:class` option, which configures
them to validate that their input is an instance of the provided class.

By default, model filters resolve the constant once and reuse it for
subsequent runs. However if `Mutations.cache_constants` is set to false,
the filter will resolve the constant before each run, which allows it to
correctly handle constants that are redefined at runtime (by Rails'
constant reloading mechanism, for example).

Array filters didn't respect the `Mutations.cache_constants` setting,
and always resolved the constant again every time the filter was called;
this brings their behaviour in line with model filters.