Retain fullname for command keys on error
The mutations gem depend on the `titleize` method in Rails that converts a string to a title like form.

E.g.: `my_app_id` to `My App Id`

Though, Pre rails `5.2`, the `titleize` method which depends on the `humanize` which truncates "_id" from the end of strings.

Example:

If I have the following command class:

```ruby

class DoStuff < Mutations::Command

  required do
    string :my_app_id
  end

  def execute
  end
end
```

And I run it, and it hard fails, the output will be:

```
Mutations::ValidationException: My App is required
```

This is a very confusing command key description, it should be `My App Id`. This PR is an attempt to handle this, and use the new Rails titleize option `keep_id_suffix: true`.

Retain fullname for command keys on error

Fix another potential bug

Remove Rails >= 5.2.0 specific test case

Use a single version of the `titleize` method and undo changes to `.gitignore`

Undo changes