Don't remove unknown characters in string filter
The `[:print:]` bracket expression only includes known characters, so it
only covers Unicode versions supported by the current version of Ruby.

This means that if a client sends a string containing a character that
wasn't yet defined when the current version of Ruby was released, it
will be incorrectly identified as "unprintable" and removed.

Instead of removing anything not included in `[:print:]`, we can remove
everything included in `[:cntrl:]`. This means unknown characters won't
be matched, and the behaviour won't depend on the current Ruby version.